Task and Appointment now extend CalendarNodeBase

+++ src/poo/calendar/IDGenerator.java
@@ -4,7 +4,7 @@ package poo.calendar;
  * Class for handling generation of unique IDs along the program
  */
 public class IDGenerator {
-	private static long mCounter = 0;
+	private static int mCounter = 0;
 	
 	// Prevent instantiation
 	private IDGenerator(){}
@@ -12,8 +12,8 @@ public class IDGenerator {
 	/**
 	 * @return a unique identifier number.
 	 */
-	public static long getID(){
-		long id = mCounter;
+	public static int getID(){
+		int id = mCounter;
 		mCounter++;
 		return id;
 	}
